requirements.txt
# python==3.11.5
charset-normalizer==3.3.2

idna==3.6
numpy==1.26.3
pvporcupine==3.0.1
requests==2.31.0
PyAudio==0.2.14

SpeechRecognition==3.10.1
typing_extensions==4.9.0
urllib3==2.1.

ffmpeg-python==0.2.0 



